+++ slurmweb/apps/gateway.py
@@ -5,6 +5,7 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import time
+import collections
 import logging
 
 from rfl.web.tokens import RFLTokenizedWebApp
@@ -23,21 +24,42 @@ from ..errors import (
 logger = logging.getLogger(__name__)
 
 
+SlurmwebAgentRacksDBSettings = collections.namedtuple(
+    "SlurmwebAgentRacksDBSettings", ["version", "infrastructure"]
+)
+
+
 class SlurmwebAgent:
-    def __init__(self, cluster, infrastructure, metrics, url):
+    def __init__(
+        self, cluster: str, racksdb: SlurmwebAgentRacksDBSettings, metrics: bool, url
+    ):
         self.cluster = cluster
-        self.infrastructure = infrastructure
         self.metrics = metrics
+        self.racksdb = racksdb
         self.url = url
 
     @classmethod
     def from_json(cls, url, data):
         try:
-            return cls(data["cluster"], data["infrastructure"], data["metrics"], url)
+            return cls(
+                data["cluster"],
+                SlurmwebAgentRacksDBSettings(**data["racksdb"]),
+                data["metrics"],
+                url,
+            )
         except KeyError as err:
             raise SlurmwebAgentError(
                 "Unable to parse cluster info fields from agent"
             ) from err
+
+
+def version_greater_or_equal(reference: str, version: str) -> bool:
+    """Return True if provided version is greater or equal than reference version."""
+
+    def version_tuple(version):
+        return tuple(int(part) for part in version.split("."))
+
+    return version_tuple(version) >= version_tuple(reference)
 
 
 class SlurmwebAppGateway(SlurmwebWebApp, RFLTokenizedWebApp):
@@ -100,6 +122,18 @@ class SlurmwebAppGateway(SlurmwebWebApp, RFLTokenizedWebApp):
                     str(err),
                 )
             else:
+                # Check version of RacksDB on agent is greater or equal than minimal
+                # version specified in configuration.
+                if not version_greater_or_equal(
+                    self.settings.agents.racksdb_version, agent.racksdb.version
+                ):
+                    logger.error(
+                        "Unsupported RacksDB API version %s on agent %s, discarding "
+                        "this agent",
+                        agent.racksdb.version,
+                        agent.cluster,
+                    )
+                    continue
                 logger.debug(
                     "Discovered available agent for cluster %s at url %s",
                     agent.cluster,